+++ e2e/tests/121-Portal/EditInfoPersonAffected/UpdateHouseNumberInvalid.spec.ts
@@ -1,6 +1,6 @@
 import HomePage from '@121-e2e/pages/Home/HomePage';
 import LoginPage from '@121-e2e/pages/Login/LoginPage';
-import RegistrationDetails from '@121-e2e/pages/RegistrationDetails/RegistrationDetailsPage';
+import PersonalInformationPopUp from '@121-e2e/pages/PersonalInformationPopUp/PersonalInformationPopUp';
 import TableModule from '@121-e2e/pages/Table/TableModule';
 import { SeedScript } from '@121-service/src/scripts/seed-script.enum';
 import NLRCProgram from '@121-service/src/seed-data/program/program-nlrc-ocw.json';
@@ -28,7 +28,7 @@ test.beforeEach(async ({ page }) => {
 test('[28044] Update HouseNumber with invalid value', async ({ page }) => {
   const table = new TableModule(page);
   const homePage = new HomePage(page);
-  const registration = new RegistrationDetails(page);
+  const piiPopUp = new PersonalInformationPopUp(page);
 
   await test.step('Navigate to PA table', async () => {
     await homePage.navigateToProgramme(NLRCProgram.titlePortal.en);
@@ -39,7 +39,7 @@ test('[28044] Update HouseNumber with invalid value', async ({ page }) => {
   });
 
   await test.step('Update house number with a string', async () => {
-    await registration.updatehousenumber({
+    await piiPopUp.updatehousenumber({
       numberString: 'house number',
     });
   });
